Compose-project/requirements.txt
Flask
mysql-connector-python
netifaces
psycopg2-binary